python_demos/requirements.txt
astroid==2.9.3
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
convertdate==2.3.2
Flask==2.0.2
hijri-converter==2.2.2
holidays==0.12
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
platformdirs==2.4.1
protobuf==3.19.3
pycodestyle==2.8.0
pylint==2.12.2
PyMeeus==0.5.11
pymongo==4.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
